backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
neo4j==5.11.0
networkx==3.1
scikit-learn==1.2.2
stix2==1.6.0
python-jose==3.3.0
python-multipart==0.0.6
pydantic==1.10.11
requests==2.31.0
